resume_aii/requirements.txt
torch
transformers
sentence-transformers
scikit-learn
pandas
numpy
uvicorn
fastapi
python-multipart
pdfplumber
python-docx
datasets
